Move types.h.
