+++ cpp/common/types.h
