+++ cpp/serving/arrow_tensor_serde.h
@@ -17,7 +17,7 @@
 #pragma once
 
 #include <serving/metaspore.pb.h>
-#include <serving/types.h>
+#include <common/types.h>
 
 #include <arrow/tensor.h>
 #include <fmt/format.h>
