+++ cpp/tests/serving/tabular_xgboost_test.cpp
@@ -21,7 +21,7 @@
 #include <serving/ort_model.h>
 #include <serving/tabular_model.h>
 #include <serving/test_utils.h>
-#include <serving/types.h>
+#include <common/types.h>
 
 #include <thread>
